ml_backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
textblob==0.17.1
nltk==3.8.1
vaderSentiment==3.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9